Screens and buttons
Boost Fingerprint  - will be on the lower right of the screen and the player can use it any time by just pressing it once to activate it. Its unlimited but not as fast as the iron boost. And player still has to dodge incoming carts. 

IRON Boost Fingerprint - The regular one will change to this when the player gets an iron boost. It only lasts 10 seconds but is faster than regular boost and the player's cart turns iron/metal. They don't have to dodge for 10 seconds and can capture all the gold.

Metal Power Up - Slowly hovering above the ground, these should show up randomly but no sooner than 1 minute apart from each other so if the player wants to get another they have to survive at least one minute. 

NAC LOGO - First thing to load on the screen 

START and HIGHSCORE Buttons: You choose where to put. The size should be minimal yet readable.